Fix(Exceptions) : All changes in sheets

+++ src/GUI/AlgeoTubes/src/main/java/besokminggu/algeotubes/AlgeoTubes.java
@@ -28,7 +28,7 @@ public class AlgeoTubes extends JFrame {
     static int ImplementasiInterpolasiBicubicSpline = 0;
     static int n;
     static int m;
-    static int x;
+    static double x;
     static String pathfile;
     
     static double[][] matriksinput;
